Add and refactor endpoints logic to manage urls

+++ src/urls/urls.service.ts
@@ -1,6 +1,6 @@
 import { ShortenUrlDto } from './dto/shorten-url.dto'
 import { PrismaService } from './../common/prisma.service'
-import { Injectable } from '@nestjs/common'
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common'
 import { createRandomString } from './helpers/create-random-string.helper'
 
 @Injectable()
@@ -10,10 +10,10 @@ export class UrlsService {
   async publicShortenUrl(shortenUrlDto: ShortenUrlDto) {
     const { longUrl } = shortenUrlDto
 
-    let shortUrl: { shortCode: string, longUrl?: string }
+    let shortUrl: { shortCode: string, longUrl: string }
 
     // verify if url already exists in database
-    const existingUrl = await this.searchUrlInDatabase(longUrl)
+    const existingUrl = await this.searchPublicUrlInDatabase(longUrl)
 
     if (!existingUrl) {
       // create random string
@@ -42,41 +42,95 @@ export class UrlsService {
   }
 
   async shortenUrl(shortenUrlDto: ShortenUrlDto, userId: number) {
-    // TODO
-    await Promise.resolve()
+    const { longUrl } = shortenUrlDto
+
+    const shortCode = await this.generateNonExistingRandomString()
+
+    const shortUrl = await this.prisma.urls.create({
+      data: {
+        longUrl,
+        shortCode,
+        createdBy: { connect: { id: userId } }
+      },
+      select: {
+        longUrl: true,
+        shortCode: true
+      }
+    })
 
     return {
-      shortenUrlDto,
-      userId
+      shortCode: shortUrl.shortCode,
+      originalUrl: shortUrl.longUrl
     }
   }
 
   async customShortenUrl(shortenUrlDto: ShortenUrlDto, userId: number) {
-    // TODO
-    await Promise.resolve()
+    const { customCode, longUrl } = shortenUrlDto
+
+    if (!customCode)
+      throw new BadRequestException('custom code was not provided')
+
+    const { exists: alreadyUsed } = await this.shortCodeExistsInDB(customCode)
+
+    if (alreadyUsed)
+      throw new BadRequestException('custom code is already in use')
+
+    const shortUrl = await this.prisma.urls.create({
+      data: {
+        longUrl,
+        shortCode: customCode,
+        createdBy: { connect: { id: userId } }
+      },
+      select: {
+        shortCode: true,
+        longUrl: true
+      }
+    })
 
     return {
-      shortenUrlDto,
-      userId
+      shortCode: shortUrl.shortCode,
+      originalUrl: shortUrl.longUrl
     }
   }
 
-  async deleteUrl(id: number) {
-    await this.prisma.urls.delete({
+  async deleteOrDisableUrl(id: number, disableOnly: boolean = true) {
+    const urlInDB = await this.prisma.urls.findFirst({
       where: { id }
     })
+
+    if (urlInDB) {
+
+      if (disableOnly) {
+        await this.prisma.urls.update({
+          where: { id },
+          data: { deleted: true }
+        })
+      } else {
+        await this.prisma.urls.delete({
+          where: { id }
+        })
+      }
+
+    } else {
+      throw new NotFoundException('url with provided id does not exists')
+    }
   }
 
-  async searchCustomCode(customCode: string) {
-    // TODO
-    await Promise.resolve()
+  async shortCodeExistsInDB(shortCode: string) {
+    const shortCodeInDb = await this.prisma.urls.findFirst({
+      where: {
+        shortCode
+      }
+    })
 
-    return customCode
+    return {
+      exists: shortCodeInDb ? true : false
+    }
   }
 
-  private async searchUrlInDatabase(longUrl: string) {
+  private async searchPublicUrlInDatabase(longUrl: string) {
     return await this.prisma.urls.findFirst({
-      where: { longUrl },
+      where: { longUrl, createdById: null },
       select: {
         shortCode: true,
         longUrl: true
@@ -87,11 +141,9 @@ export class UrlsService {
   private async generateNonExistingRandomString(): Promise<string> {
     const randomString = createRandomString(6)
 
-    const randomStringExistsInDb = await this.prisma.urls.findFirst({
-      where: { shortCode: randomString }
-    })
+    const { exists } = await this.shortCodeExistsInDB(randomString)
 
-    if (!randomStringExistsInDb) return randomString
+    if (!exists) return randomString
     else return await this.generateNonExistingRandomString()
   }
 }
